+++ max/tests/integration/architectures/deepseekV2/test_latent_attention_fp8.py
@@ -233,7 +233,6 @@ def generate_max_outputs_fp8(
         n_kv_heads=1,
         head_dim=576,
         num_layers=config.num_hidden_layers,
-        cache_strategy="paged",
         devices=[DeviceRef.GPU()],
         page_size=128,
         is_mla=True,